Modify exports and pass resources to durable execute
