+++ index.ts
@@ -1,9 +1,11 @@
-export { registerActivities } from "./activities";
+export { _registerActivities } from "./activities";
 import { appendOutput, setOutput } from "./output";
 import sql from "./sql";
 import { execute } from "./tasks";
-export { ExecuteOptions } from "./tasks";
+export { ExecuteOptions } from "./api";
 export * as resources from "./resources";
+
+export { appendOutput, setOutput, execute, sql };
 
 export default {
   appendOutput,
